Split helper test, move fuzzing to api package as serialization_test.go
